start refactoring server transport
